feat(PRL-5788): Integrate dashboard notification closure feature
