+++ src/main/steps/applicant/confirm-contact-details/address/lookup/content.test.ts
@@ -1,3 +1,4 @@
+import { ResourceReader } from '../.././../../../modules/resourcereader/ResourceReader';
 import { FormFields } from '../../../../../app/form/Form';
 import { CommonContent } from '../../../../common/common.content';
 import {
@@ -5,7 +6,6 @@ import {
   generateContent as generateAddressLookupContent,
 } from '../../../../common/components/address-lookup';
 import { APPLICANT_MANUAL_ADDRESS } from '../../../../urls';
-import { ResourceReader } from '../.././../../../modules/resourcereader/ResourceReader';
 
 import { generateContent } from './content';
