+++ src/main/steps/c100-rebuild/check-your-answers/util/safetyConcerns.util.test.ts
@@ -10,13 +10,13 @@ import {
   en as passportAmountEn,
 } from '../../../common/safety-concerns/abduction/passport-amount/content';
 import {
-  cy as previousAbductionCy,
-  en as previousAbductionEn,
-} from '../../../common/safety-concerns/abduction/passport-office-notified/content';
-import {
   cy as passportOfficeNotifiedCy,
   en as passportOfficeNotifiedEn,
 } from '../../../common/safety-concerns/abduction/passport-office/content';
+import {
+  cy as previousAbductionCy,
+  en as previousAbductionEn,
+} from '../../../common/safety-concerns/abduction/passport-office-notified/content';
 import {
   cy as passportOfficeCy,
   en as passportOfficeEn,
