+++ src/main/steps/common/safety-concerns/review/util/safetyConcerns.util.ts
@@ -3,11 +3,11 @@
 
 import { cy as abductionChildLocationCy, en as abductionChildLocationEn } from '../../abduction/child-location/content';
 import { cy as passportAmountCy, en as passportAmountEn } from '../../abduction/passport-amount/content';
+import { cy as passportOfficeCy, en as passportOfficeEn } from '../../abduction/passport-office/content';
 import {
   cy as passportOfficeNotifiedCy,
   en as passportOfficeNotifiedEn,
 } from '../../abduction/passport-office-notified/content';
-import { cy as passportOfficeCy, en as passportOfficeEn } from '../../abduction/passport-office/content';
 import { cy as previousAbductionCy, en as previousAbductionEn } from '../../abduction/previousabductions/content';
 import { cy as abducionThreatsCy, en as abducionThreatsEn } from '../../abduction/threats/content';
 import { cy as concernForChildSafetyCy, en as concernForChildSafetyEn } from '../../child/concerns-about/content';
